# server/openapi.yaml
openapi: "3.0.3"
info:
  title: "GIAM-SAT API v1.13.0"
  description: |
    ## GIAM-SAT Security Monitoring System

    REST API for the GIAM-SAT security monitoring platform.
    
    **Features:**
    - Agent monitoring & management
    - Event log collection & analysis  
    - File integrity monitoring (FIM)
    - Network traffic capture (Wireshark-level)
    - Vulnerability scanning (NVD + CVE)
    - YARA malware detection
    - SCA/CIS compliance assessment
    - Threat intelligence (OTX + VirusTotal + URLhaus + MISP)
    - AI-powered security analysis (DeepSeek/OpenAI/Gemini/Groq/xAI)
    - Telegram Bot notifications
    - Agentless monitoring (SNMP/SSH/Ping)
    
    **Authentication:** JWT Bearer token (except localhost)
    
    **Rate Limits:** 5 req/min for `/api/float_ai_chat`
    
    **Base URL:** `http://{host}:5000`
    
  version: "1.13.0"
  contact:
    name: "GIAM-SAT"
  license:
    name: "Proprietary"
servers:
  - url: "http://localhost:5000"
    description: "Local development server"
  - url: "http://{server_ip}:5000"
    description: "Production server"
    variables:
      server_ip:
        default: "YOUR_SERVER_IP"
        description: "GIAM-SAT server IP address"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token from /api/auth/login"
  schemas:
...
    Machine:
      type: object
      properties:
        machine_id:
          type: string
          example: "DESKTOP-ABC123"
        hostname:
          type: string
          example: "DESKTOP-ABC123"
        ip:
          type: string
          example: "192.168.1.100"
        os:
          type: string
          example: "Windows 11 Pro"
        online:
          type: boolean
          example: true
        last_seen:
          type: string
          format: date-time
        cpu_percent:
          type: number
          example: 45.2
        memory_percent:
          type: number
          example: 62.7
        disk_percent:
          type: number
          example: 55.0
    Event:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        event_id:
          type: string
          example: "4625"
        channel:
          type: string
          example: "Security"
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
    NetworkPacket:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        src_ip:
          type: string
          example: "192.168.1.100"
        dst_ip:
          type: string
          example: "8.8.8.8"
        src_port:
          type: integer
          example: 54321
        dst_port:
          type: integer
          example: 53
        protocol:
          type: string
          example: "UDP"
        protocol_app:
          type: string
          example: "DNS"
        dns_info:
          type: object
          properties:
            qname:
              type: string
            qtype_name:
              type: string
            rcode_name:
              type: string
        http_info:
          type: object
          properties:
            http_method:
              type: string
            http_uri:
              type: string
        raw_data:
          type: string
          description: "Full packet JSON with hex dump"
        timestamp:
          type: string
          format: date-time
    ThreatAlert:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        rule_id:
          type: string
          example: "THREAT-001"
        rule_name:
          type: string
          example: "Brute Force Attack Detected"
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        description:
          type: string
        mitre:
          type: string
          example: "T1110"
        timestamp:
          type: string
          format: date-time
    VulnerabilityAlert:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        cve_id:
          type: string
          example: "CVE-2024-1234"
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        description:
          type: string
        cvss_score:
          type: number
          example: 9.8
        timestamp:
          type: string
          format: date-time
    YaraAlert:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        rule_name:
          type: string
          example: "MAL_ransomware_behavior"
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        description:
          type: string
        file_path:
          type: string
        timestamp:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/machines:
    get:
      tags: [Machines]
      summary: "List all agent machines"
      parameters:
        - name: online_only
          in: query
          schema:
            type: boolean
          description: "Filter online machines only"
      responses:
        "200":
          description: "List of machines"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
...
  /api/events:
    get:
      tags: [Events]
      summary: "Get event logs"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
          description: "Filter by machine"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: "Max results"
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        - name: event_id
          in: query
          schema:
            type: string
          description: "Filter by Windows Event ID"
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: "Events after this timestamp"
      responses:
        "200":
          description: "List of events"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
...
  /api/network:
    get:
      tags: [Network]
      summary: "Get network traffic logs with application layer data"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: protocol
          in: query
          schema:
            type: string
            enum: [TCP, UDP, ICMP]
        - name: protocol_app
          in: query
          schema:
            type: string
            enum: [DNS, HTTP, TLS]
          description: "Filter by application layer protocol"
      responses:
        "200":
          description: "List of network packets with application layer info"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NetworkPacket"
...
  /api/threats:
    get:
      tags: [Threats]
      summary: "Get threat/correlation alerts"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
      responses:
        "200":
          description: "List of threat alerts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatAlert"
...
  /api/vulns:
    get:
      tags: [Vulnerabilities]
      summary: "Get CVE vulnerability alerts"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
      responses:
        "200":
          description: "List of vulnerability alerts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VulnerabilityAlert"
...
  /api/yara:
    get:
      tags: [YARA]
      summary: "Get YARA malware detection alerts"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: "List of YARA alerts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YaraAlert"
...